audio.c: added debugging lines as well as bug fixes in the pitch detector
audio.h: changed the numerator and denominator of the ratio struct to be floats

extras/cli_phase_vocoder/Makefile: changed the test makefile to include audio.c/h

extras/cli_phase_vocoder/Makefile: added tests for pitch detecting

global.h: added a #define for debug output

added sweep.wav in the samples folder of the repo

+++ audio.h
@@ -29,8 +29,8 @@ struct audio_out_t
 
 typedef struct
 {
-    int numerator; //new frequency
-    int denominator; //old frequency
+    float numerator; //new frequency
+    float denominator; //old frequency
 } freq_ratio;
 
 #ifdef FOR_DSP
